move auth expected_keepers to internal anteInterface
