+++ x/feeshare/ante/ante.go
@@ -3,21 +3,20 @@ package ante
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 
 	wasmtypes "github.com/CosmWasm/wasmd/x/wasm/types"
-	authforktypes "github.com/CosmosContracts/juno/v12/x/auth/types"
 	feeshare "github.com/CosmosContracts/juno/v12/x/feeshare/types"
-	"github.com/cosmos/cosmos-sdk/x/auth/types"
 )
 
 // FeeSharePayoutDecorator Run his after we already deduct the fee from the account with
 // the ante.NewDeductFeeDecorator() decorator. We pull funds from the FeeCollector ModuleAccount
 type FeeSharePayoutDecorator struct {
-	bankKeeper     authforktypes.BankKeeper
-	feesharekeeper authforktypes.FeeShareKeeper
+	bankKeeper     BankKeeper
+	feesharekeeper FeeShareKeeper
 }
 
-func NewFeeSharePayoutDecorator(bk authforktypes.BankKeeper, fs authforktypes.FeeShareKeeper) FeeSharePayoutDecorator {
+func NewFeeSharePayoutDecorator(bk BankKeeper, fs FeeShareKeeper) FeeSharePayoutDecorator {
 	return FeeSharePayoutDecorator{
 		bankKeeper:     bk,
 		feesharekeeper: fs,
@@ -55,7 +54,7 @@ func FeePayLogic(fees sdk.Coins, govPercent sdk.Dec, numPairs int) sdk.Coins {
 
 // FeeSharePayout takes the total fees and redistributes 50% (or param set) to the contract developers
 // provided they opted-in to payments.
-func FeeSharePayout(ctx sdk.Context, bankKeeper authforktypes.BankKeeper, totalFees sdk.Coins, revKeeper authforktypes.FeeShareKeeper, msgs []sdk.Msg) error {
+func FeeSharePayout(ctx sdk.Context, bankKeeper BankKeeper, totalFees sdk.Coins, revKeeper FeeShareKeeper, msgs []sdk.Msg) error {
 
 	params := revKeeper.GetParams(ctx)
 	if !params.EnableFeeShare {
@@ -104,7 +103,7 @@ func FeeSharePayout(ctx sdk.Context, bankKeeper authforktypes.BankKeeper, totalF
 
 		// pay fees evenly between all withdraw addresses
 		for _, withdrawAddr := range toPay {
-			err := bankKeeper.SendCoinsFromModuleToAccount(ctx, types.FeeCollectorName, withdrawAddr, splitFees)
+			err := bankKeeper.SendCoinsFromModuleToAccount(ctx, authtypes.FeeCollectorName, withdrawAddr, splitFees)
 			if err != nil {
 				return sdkerrors.Wrapf(feeshare.ErrFeeSharePayment, "failed to pay fees to contract developer: %s", err.Error())
 			}
