+++ x/feeshare/ante/expected_keepers.go
@@ -1,4 +1,4 @@
-package types
+package ante
 
 // Used for the Juno ante handler so we can properly send 50% of fees to dAPP developers via fee share module
